feat: all subjects
added the main subjects and their exercises (solutions are
stored separately), as well as
their (not double checked) docs. Macro's in Game seem
to be a lot of fun, induction was rather tough.

The types (inductive, structures) could be improved.
Need to add a guideline, how to go through the exercises.